New Gear Ratio!
Originally SmartSpin2K was designed for Magnetic resistance bikes but since most users are actually using friction resistance bikes we need a bit more torque to turn the knobs. This new gear ratio fits in the place of the old one (some sanding on the gear shroud required) and increases the torque by nearly 33%. while still keeping torque low enough as to not break anything on the SS2k or your bike.