Add files via upload
Hello All,

Doudar originaly mentioned this as a possibility for the SS2KND (Narrow Design). This suggestion is to move the circuit for resisitor "up" and eliminate the pin holes. The case would then have a fence printed (or milled) at the bottom similar to the stepper motor (lid) fence, which would allow the PCB to be "plugged in" (glue optional).

-keith