Added single sided rigid mounting strap
The idea of this strap is for the front to pass by the head tube and then get secured (lashed) to the head tube with velcro, rubber bands, rope, tape, etc.

This might make a more stable mount for installations that have a long head tube to SS2K distance.

Co-Authored-By: Nick B <76080977+fireyeti@users.noreply.github.com>
Co-Authored-By: Joel Baranick <kadaan@users.noreply.github.com>
Co-Authored-By: jkbaker4 <40582757+jkbaker4@users.noreply.github.com>