Adding parameters for auto-home
Added parameters for auto homing.
minStep is the minimum we will allow the stepper to travel.
maxStep is the maximum we will allow the stepper to travel.
maxWatts is the maximum watts the brake on the spin bike can absorb from the user.

Theory is that minStep will be automatically set by finding 50w on the Power Table.
maxStep will be set by automatically set by calculating the position based on the watt table and maxWatts, which will be user defined.